backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
redis==5.0.1
ray==2.9.0
rdkit-pypi==2023.9.4
selfies==2.1.1
umap-learn==0.5.5
numpy==1.26.2
scipy==1.11.4
scikit-learn==1.3.2
aiofiles==23.2.1
pydantic==2.5.2
vina==1.2.3
